requirements.txt
Django==1.9.5
django-ckeditor==5.0.3
django-mptt==0.8.3
django-mptt-admin==0.3.6
django-suit==0.2.18
django-tinymce==2.3.0
Pillow==3.2.0
psycopg2==2.6.1
six==1.10.0
